refactor(PR-2/3): 탭/모달 컴포넌트 분리 및 BenefitsEngine 완전 통합
PR-2 보완: App.jsx 988줄 달성 (2475줄 → 988줄, 60% 감소)
- HomeTab, BenefitsTab, WalletTab, SettingsTab 컴포넌트 분리
- PlaceSheet 바텀시트 컴포넌트 분리
- OcrModal 컴포넌트 분리
- 사용하지 않는 import 정리

PR-3 보완: BenefitsEngine 완전 통합
- cardRanking을 engine.calculateRanking()으로 교체
- 중복된 랭킹 계산 로직 제거
- O(1) 인덱스 기반 조회로 성능 최적화

검증:
- ESLint: 0 errors, 0 warnings
- Build: 성공
