+++ src/components/index.js
@@ -7,3 +7,5 @@ export { LoadingScreen } from './LoadingScreen';
 export { ErrorScreen } from './ErrorScreen';
 export { BenefitDetailModal } from './BenefitDetailModal';
 export { MapView } from './MapView';
+export { PlaceSheet } from './PlaceSheet';
+export { OcrModal } from './OcrModal';
