+++ src/tabs/index.js
@@ -1,0 +1,8 @@
+/**
+ * Tabs index - re-export all tab components
+ */
+
+export { HomeTab } from './HomeTab';
+export { BenefitsTab } from './BenefitsTab';
+export { WalletTab } from './WalletTab';
+export { SettingsTab } from './SettingsTab';
